# Conflicts:
#	src/main/java/school/faang/user_service/mapper/UserScoreMapper.java

+++ src/main/java/school/faang/user_service/leaderboard/ScoreCalculator.java
@@ -5,5 +5,5 @@ import school.faang.user_service.dto.user.UserScoreDto;
 
 public interface ScoreCalculator {
 
-    long getScore(@NotNull UserScoreDto userDto);
+    int getScore(@NotNull UserScoreDto userDto);
 }
